Improve main menu generator
- allow any dynamic items to be added
- rewrite submenus as data classes
