+++ locale/szl/LC_MESSAGES/kontext.po
@@ -459,3 +459,6 @@ msgstr "Concordance regeneration"
 
 msgid "Paradigmatic query calculation"
 msgstr "Paradigmatic query calculation UNTRANSLATED"
+
+msgid "Go to the constituent concordance #{}"
+msgstr "Go to the constituent concordance #{} UNTRANSLATED"
